requirements.txt
pytest==7.1.1
Appium-Python-Client==1.1.0
pytest==7.1.1
selenium==3.141.0
webdriver-manager==3.5.4
pytest-xdist==2.5.0
allure-pytest==2.8.32
allure-python-commons==2.8.32
